Correct: C 1045

+++ C/1045 [기초-산술연산] 정수 2개 입력받아 자동 계산하기.c
@@ -9,7 +9,7 @@ int main(void) {
 	printf("%lld\n", num1 * num2);
 	printf("%lld\n", num1 / num2);
 	printf("%lld\n", num1 % num2);
-	printf("%.2f\n", num1 / num2);
+	printf("%.2f\n", (float)num1 / num2); // num1을 folat으로 형 변환 
 	
 	return 0;
 }
